Separate script/sign
