+++ src/scriptutils.h
@@ -8,17 +8,8 @@
 
 #include "key.h"
 #include "script/script.h"
-#include "script/interpreter.h"
-#include "script/standard.h"
-
-#include <stdexcept>
-#include <stdint.h>
-#include <string>
-#include <vector>
 
 class CKeyStore;
-class CTransaction;
-struct CMutableTransaction;
 
 /** IsMine() return codes */
 enum isminetype
@@ -34,11 +25,5 @@ typedef uint8_t isminefilter;
 isminetype IsMine(const CKeyStore& keystore, const CScript& scriptPubKey);
 isminetype IsMine(const CKeyStore& keystore, const CTxDestination& dest);
 void ExtractAffectedKeys(const CKeyStore &keystore, const CScript& scriptPubKey, std::vector<CKeyID> &vKeys);
-bool SignSignature(const CKeyStore& keystore, const CScript& fromPubKey, CMutableTransaction& txTo, unsigned int nIn, int nHashType=SIGHASH_ALL);
-bool SignSignature(const CKeyStore& keystore, const CTransaction& txFrom, CMutableTransaction& txTo, unsigned int nIn, int nHashType=SIGHASH_ALL);
-
-// Given two sets of signatures for scriptPubKey, possibly with OP_0 placeholders,
-// combine them intelligently and return the result.
-CScript CombineSignatures(CScript scriptPubKey, const CTransaction& txTo, unsigned int nIn, const CScript& scriptSig1, const CScript& scriptSig2);
 
 #endif // H_BITCOIN_SCRIPT
